# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/driver": {
            "get": {
                "description": "Get a list of Driver.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Get All Driver.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Driver"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Driver.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Create New Driver.",
                "parameters": [
                    {
                        "description": "The body to create a new Driver",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.driverInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Driver"
                        }
                    }
                }
            }
        },
        "/driver/{id}": {
            "get": {
                "description": "Get a Driver by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Get Driver by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Driver"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Driver by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Delete one driver.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Driver by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Update Driver.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The body to update an Driver",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.driverInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Driver"
                        }
                    }
                }
            }
        },
        "/driverincentive": {
            "get": {
                "description": "Get a list of Driver Incentive.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver Incentive"
                ],
                "summary": "Get All Driver Incentive.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DriverIncentive"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Driver Incentive.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver Incentive"
                ],
                "summary": "Create New Driver Incentive.",
                "parameters": [
                    {
                        "description": "The body to create a new Driver Incentive",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.driverIncentiveInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DriverIncentive"
                        }
                    }
                }
            }
        },
        "/driverincentive/{id}": {
            "get": {
                "description": "Get a Driver by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver Incentive"
                ],
                "summary": "Get Driver by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Driver"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Driver Incentive by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver Incentive"
                ],
                "summary": "Delete one Driver Incentive.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DriverIncentive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Driver Incentive by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver Incentive"
                ],
                "summary": "Update Driver Incentive.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DriverIncentive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The body to update an Driver Incentive",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.driverIncentiveInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DriverIncentive"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.driverIncentiveInput": {
            "type": "object",
            "required": [
                "booking_id",
                "incentive"
            ],
            "properties": {
                "booking_id": {
                    "type": "integer"
                },
                "incentive": {
                    "type": "integer"
                }
            }
        },
        "controllers.driverInput": {
            "type": "object",
            "required": [
                "daily_cost",
                "name",
                "nik",
                "phone_number"
            ],
            "properties": {
                "daily_cost": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "models.Booking": {
            "type": "object",
            "required": [
                "booking_type_id",
                "cars_id",
                "customer_id",
                "driver_id",
                "end_rent",
                "start_rent",
                "total_cost"
            ],
            "properties": {
                "booking_type": {
                    "$ref": "#/definitions/models.BookingType"
                },
                "booking_type_id": {
                    "type": "integer"
                },
                "cars": {
                    "$ref": "#/definitions/models.Cars"
                },
                "cars_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/models.Customer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "discount": {
                    "type": "integer"
                },
                "driver": {
                    "$ref": "#/definitions/models.Driver"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_rent": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "start_rent": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "integer"
                },
                "total_driver_cost": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BookingType": {
            "type": "object",
            "required": [
                "booking_type",
                "description"
            ],
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Cars": {
            "type": "object",
            "required": [
                "daily_rent",
                "name",
                "stock"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_rent": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "required": [
                "name",
                "nik",
                "phone_number"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "membership": {
                    "$ref": "#/definitions/models.Membership"
                },
                "membership_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Driver": {
            "type": "object",
            "required": [
                "name",
                "nik",
                "phone_number"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_cost": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.DriverIncentive": {
            "type": "object",
            "required": [
                "booking_id",
                "incentive"
            ],
            "properties": {
                "booking": {
                    "$ref": "#/definitions/models.Booking"
                },
                "booking_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incentive": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Membership": {
            "type": "object",
            "required": [
                "discount",
                "membership_name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "membership_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}